% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM-Analysis.R
\name{DTM.density}
\alias{DTM.density}
\title{Estimation of DTM-Densities}
\usage{
DTM.density(
  sample,
  m,
  r = 2,
  kernel = "biweight",
  bw = "SJ",
  full_return = F,
  ...
)
}
\arguments{
\item{sample}{either a n x d numeric coordinate matrix  or a list of such matrices. Here, n corresponds to the size of the sample and d to the dimension of the data.}

\item{m}{the mass parameter of the corresponding (empirical) DTM-functions. It is a double in (0,1].}

\item{r}{a double in the definition of the general (empirical) DTM-function. By definition \code{r >= 1} and the default is \code{r = 2}.}

\item{kernel}{a character string giving the smoothing kernel to be used. This must partially match one of "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine" or "optcosine", with default "biweight".}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that this is the standard deviation of the smoothing kernel. bw can also be a character string giving a rule to choose the bandwidth. See \code{\link[stats]{bw.nrd}}. By default \code{bw = "SJ"}.}

\item{full_return}{a boolean specifying whether the output of the function should also include the DTM-function evaluated at each point of the input sample (disabled by default).}

\item{...}{additional arguments passed to the function \code{\link[stats]{density}}.}
}
\value{
Let \code{sample} be a n x d matrix. Then, the function returns an estimator for the DTM-density proposed in \insertCite{proksch2022small;textual}{DTMdemo} based the given sample (i.e., an object of class "\code{\link[stats]{density}}"). If \code{full_return} is \code{TRUE}, a list with two components will be returned. It contains the the DTM-function evaluated at each sample point ("\code{dtm}") and the corresponding kernel density estimator ("\code{kde}").

If \code{sample} is a list of input matrices, then a list containing the return of each input matrix is returned.
}
\description{
Calculate the DTM-densities estimators proposed in \insertCite{proksch2022small;textual}{DTMdemo} for all input samples.
}
\details{
Given a n x d coordinate matrix, the (r'th power of the) empirical DTM-function is calculated at each sample point encoded in said matrix. To this end, the function \code{\link[TDA]{dtm}} is used. The resulting vector is used to estimate the DTM-density of the input sample via a kernel density estimator (see \insertCite{proksch2022small;textual}{DTMdemo} for details).

 If the input is a list of coordinate matrices, this procedure is repeated for every entry of the list.
}
\examples{
n= 1000

square = matrix(0,n,2)
square[,1] = runif(n,0,1)
square[,2] = runif(n,0,1)

disc = matrix(0,n,2)
radius = 0.5
r = radius*sqrt(runif(n,0,1))
theta = runif(n,0,1)*2*pi
disc[,1]= r*cos(theta)
disc[,2]= r*sin(theta)

sample = list(square, disc)

den.list=DTM.density(sample, m=1)
plot(den.list[[1]], ylim =c(0,5), main = "")
lines(den.list[[2]], col = "red")
}
\references{
{
  \insertAllCited{}
}
}
