% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM-Analysis.R
\name{DTM.function}
\alias{DTM.function}
\title{Calculation of the Empirical DTM-Function}
\usage{
DTM.function(sample, m, r = 2)
}
\arguments{
\item{sample}{either a n x d numeric coordinate matrix  or a list of such matrices. Here, n corresponds to the size of the sample and d to the dimension of the data.}

\item{m}{the mass parameter of the (empirical) DTM-function. It is a double in (0,1].}

\item{r}{a double in the definition of the general (empirical) DTM-function. By definition r>=1 and the default value is 2.}
}
\value{
Let \code{sample} be a n x d matrix. Then, the function returns a vector that contains the value of the (r'th power of the) empirical DTM-function evaluated at each point encoded in \code{sample}.

If \code{sample} is a list of input matrices, then a list containing the return of each input matrix is returned.
}
\description{
This function calculates the (r'th power of the) empirical DTM-function (see \insertCite{chazal2016rates;textual}{DTMdemo} for a definition) for all points in a given sample.  It is a wrapper for \code{\link[TDA]{dtm}}.
}
\examples{
n <- 1000

square <- matrix(0,n,2)
square[,1] <- runif(n,0,1)
square[,2] <- runif(n,0,1)

disc <- matrix(0,n,2)
radius <- 0.5
r <- radius*sqrt(runif(n,0,1))
theta <- runif(n,0,1)*2*pi
disc[,1] <- r*cos(theta)
disc[,2] <- r*sin(theta)

sample <- list(square, disc)

DTM.function(sample, m = 1)
}
\references{
{
  \insertAllCited{}
}
}
