% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTM-Analysis.R
\name{L1.density.dist}
\alias{L1.density.dist}
\title{L1-distance between Kernel Density Estimators}
\usage{
L1.density.dist(density.list, lower, upper, ...)
}
\arguments{
\item{density.list}{a list of objects with class "\code{\link[stats]{density}}".}

\item{lower, upper}{the limits of integration. Can be infinite.}

\item{...}{additional arguments passed to the function "\code{\link[stats]{integrate}}."}
}
\value{
The function returns an estimate for the absolute integrated difference matrix for the elements in \code{density.list} (integration from \code{lower} to \code{upper}).
}
\description{
Calculate the L1-distance distance matrix for a list of objects with class "\code{\link[stats]{density}}".
}
\examples{
n = 1000

square <- matrix(0,n,2)
square[,1] <- runif(n,0,1)
square[,2] <- runif(n,0,1)

disc <- matrix(0,n,2)
radius <- 0.5
r <- radius*sqrt(runif(n,0,1))
theta <- runif(n,0,1)*2*pi
disc[,1] <- r*cos(theta)
disc[,2] <- r*sin(theta)

sample <- list(square, disc)

den.list <- DTM.density(sample, m = 1)
L1.density.dist(den.list, 0, 1)
}
